refactor(layout): unify landing and authenticated nav bars (#106)
Consolidates the two parallel nav implementations (inline landing nav
in app/page.tsx and components/layout/nav-header.tsx) into a single
NavBar component that adapts to auth state via an authState prop.

- components/layout/nav-bar.tsx: unified responsive nav with
  authenticated and unauthenticated variants, shared logo, brand
  colors, mobile hamburger, and sign-out flow.
- app/page.tsx: renders <NavBar authState="unauthenticated" />.
- app/(app)/layout.tsx: renders <NavBar authState="authenticated" />.
- Removes nav-header.tsx and its test, replaced by nav-bar.test.tsx
  which covers both auth states plus mobile menu behavior.
- landing-page.test.tsx: adds next/navigation hook and supabase client
  mocks now that the landing page embeds the client NavBar.

No behavior change for end users; the visible links, routes, and
sign-out flow are preserved exactly.
