+++ app/(app)/layout.tsx
@@ -1,15 +1,16 @@
-import { NavHeader } from "@/components/layout";
+import { NavBar } from "@/components/layout";
 
 /**
  * Layout for authenticated app pages.
  *
- * Wraps all routes under `(app)/` with the navigation header.
- * Individual pages handle their own auth checks and redirects.
+ * Wraps all routes under `(app)/` with the unified navigation bar in
+ * its authenticated variant. Individual pages handle their own auth
+ * checks and redirects.
  */
 export default function AppLayout({ children }: { children: React.ReactNode }) {
   return (
     <>
-      <NavHeader />
+      <NavBar authState="authenticated" />
       {children}
     </>
   );
