+++ components/layout/index.ts
@@ -1,3 +1,4 @@
-export { NavHeader } from "./nav-header";
+export { NavBar } from "./nav-bar";
+export type { NavBarAuthState, NavBarProps } from "./nav-bar";
 export { PageContainer } from "./page-container";
 export type { PageContainerProps, ContainerWidth } from "./page-container";
